test: exract user task test helpers
Creates a `ZeebeResourcesHelper` utility for integration tests. It uses a defined
Zeebe client to create resources in Zeebe, e.g. process instances. The helpers are
extracted from the `GrpcClientRule` and the `UserTaskTest` into a common utility
that is reused in the `GrpcClientRule` and the user task integration tests.

Removes the `UserTaskTest` class since it's obsolete now. The user task tests classes
handle their setup themselves.
