+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/client/command/AssignUserTaskTest.java
@@ -10,14 +10,48 @@ package io.camunda.zeebe.it.client.command;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import io.camunda.zeebe.client.ZeebeClient;
 import io.camunda.zeebe.client.api.command.ProblemException;
 import io.camunda.zeebe.it.util.ZeebeAssertHelper;
+import io.camunda.zeebe.it.util.ZeebeResourcesHelper;
+import io.camunda.zeebe.qa.util.cluster.TestCluster;
+import io.camunda.zeebe.qa.util.cluster.TestZeebePort;
+import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
+import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
+import io.camunda.zeebe.test.util.junit.AutoCloseResources;
+import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import java.time.Duration;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public final class AssignUserTaskTest extends UserTaskTest {
+@ZeebeIntegration
+@AutoCloseResources
+class AssignUserTaskTest {
+
+  @TestZeebe(clusterSize = 1, partitionCount = 1, replicationFactor = 1)
+  static final TestCluster CLUSTER =
+      TestCluster.builder().withEmbeddedGateway(false).withGatewaysCount(1).build();
+
+  @AutoCloseResource ZeebeClient client;
+
+  private long userTaskKey;
+
+  @BeforeEach
+  void initClientAndInstances() {
+    final var gateway = CLUSTER.availableGateway();
+    client =
+        CLUSTER
+            .newClientBuilder()
+            .gatewayAddress(gateway.gatewayAddress())
+            .gatewayRestApiPort(gateway.mappedPort(TestZeebePort.REST))
+            .defaultRequestTimeout(Duration.ofSeconds(15))
+            .build();
+    final ZeebeResourcesHelper resourcesHelper = new ZeebeResourcesHelper(client);
+    userTaskKey = resourcesHelper.createSingleUserTask();
+  }
 
   @Test
-  public void shouldAssignUserTaskWithAssignee() {
+  void shouldAssignUserTaskWithAssignee() {
     // when
     client.newUserTaskAssignCommand(userTaskKey).assignee("barbar").send().join();
 
@@ -31,7 +65,7 @@ public final class AssignUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldAssignUserTaskWithAssigneeWithImplicitAllowOverride() {
+  void shouldAssignUserTaskWithAssigneeWithImplicitAllowOverride() {
     // given
     client.newUserTaskAssignCommand(userTaskKey).assignee("foobar").send().join();
 
@@ -44,7 +78,7 @@ public final class AssignUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldAssignUserTaskWithAssigneeWithExplicitAllowOverride() {
+  void shouldAssignUserTaskWithAssigneeWithExplicitAllowOverride() {
     // given
     client.newUserTaskAssignCommand(userTaskKey).assignee("foobar").send().join();
 
@@ -62,7 +96,7 @@ public final class AssignUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldAssignUserTaskWithAssigneeAndAction() {
+  void shouldAssignUserTaskWithAssigneeAndAction() {
     // when
     client.newUserTaskAssignCommand(userTaskKey).assignee("barbar").action("foo").send().join();
 
@@ -76,7 +110,7 @@ public final class AssignUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldRejectIfUserTaskIsAlreadyAssignedWithProhibitedOverride() {
+  void shouldRejectIfUserTaskIsAlreadyAssignedWithProhibitedOverride() {
     // given
     client.newUserTaskAssignCommand(userTaskKey).assignee("foobar").send().join();
 
@@ -94,7 +128,7 @@ public final class AssignUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldRejectIfMissingAssignee() {
+  void shouldRejectIfMissingAssignee() {
     // when / then
     assertThatThrownBy(() -> client.newUserTaskAssignCommand(userTaskKey).send().join())
         .hasCauseInstanceOf(ProblemException.class)
@@ -102,7 +136,7 @@ public final class AssignUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldRejectIfMissingAssigneeWithProhibitedOverride() {
+  void shouldRejectIfMissingAssigneeWithProhibitedOverride() {
     // when / then
     assertThatThrownBy(
             () -> client.newUserTaskAssignCommand(userTaskKey).allowOverride(false).send().join())
