+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/client/command/CompleteUserTaskTest.java
@@ -11,15 +11,49 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.entry;
 
+import io.camunda.zeebe.client.ZeebeClient;
 import io.camunda.zeebe.client.api.command.ProblemException;
 import io.camunda.zeebe.it.util.ZeebeAssertHelper;
+import io.camunda.zeebe.it.util.ZeebeResourcesHelper;
+import io.camunda.zeebe.qa.util.cluster.TestCluster;
+import io.camunda.zeebe.qa.util.cluster.TestZeebePort;
+import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
+import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
+import io.camunda.zeebe.test.util.junit.AutoCloseResources;
+import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import java.time.Duration;
 import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public final class CompleteUserTaskTest extends UserTaskTest {
+@ZeebeIntegration
+@AutoCloseResources
+class CompleteUserTaskTest {
+
+  @TestZeebe(clusterSize = 1, partitionCount = 1, replicationFactor = 1)
+  private static final TestCluster CLUSTER =
+      TestCluster.builder().withEmbeddedGateway(false).withGatewaysCount(1).build();
+
+  @AutoCloseResource private ZeebeClient client;
+
+  private long userTaskKey;
+
+  @BeforeEach
+  void initClientAndInstances() {
+    final var gateway = CLUSTER.availableGateway();
+    client =
+        CLUSTER
+            .newClientBuilder()
+            .gatewayAddress(gateway.gatewayAddress())
+            .gatewayRestApiPort(gateway.mappedPort(TestZeebePort.REST))
+            .defaultRequestTimeout(Duration.ofSeconds(15))
+            .build();
+    final ZeebeResourcesHelper resourcesHelper = new ZeebeResourcesHelper(client);
+    userTaskKey = resourcesHelper.createSingleUserTask();
+  }
 
   @Test
-  public void shouldCompleteUserTask() {
+  void shouldCompleteUserTask() {
     // when
     client.newUserTaskCompleteCommand(userTaskKey).send().join();
 
@@ -33,7 +67,7 @@ public final class CompleteUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldCompleteUserTaskWithAction() {
+  void shouldCompleteUserTaskWithAction() {
     // when
     client.newUserTaskCompleteCommand(userTaskKey).action("foo").send().join();
 
@@ -47,7 +81,7 @@ public final class CompleteUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldCompleteUserTaskWithVariables() {
+  void shouldCompleteUserTaskWithVariables() {
     // when
 
     client.newUserTaskCompleteCommand(userTaskKey).variables(Map.of("foo", "bar")).send().join();
@@ -59,7 +93,7 @@ public final class CompleteUserTaskTest extends UserTaskTest {
   }
 
   @Test
-  public void shouldRejectIfJobIsAlreadyCompleted() {
+  void shouldRejectIfJobIsAlreadyCompleted() {
     // given
     client.newUserTaskCompleteCommand(userTaskKey).send().join();
