+++ zeebe/qa/integration-tests/src/test/java/io/camunda/zeebe/it/util/GrpcClientRule.java
@@ -7,31 +7,17 @@
  */
 package io.camunda.zeebe.it.util;
 
-import static io.camunda.zeebe.test.util.TestUtil.waitUntil;
-import static org.assertj.core.api.Assertions.assertThat;
-
 import io.camunda.zeebe.broker.TestLoggers;
 import io.camunda.zeebe.broker.test.EmbeddedBrokerRule;
 import io.camunda.zeebe.client.ZeebeClient;
 import io.camunda.zeebe.client.ZeebeClientBuilder;
-import io.camunda.zeebe.client.api.response.DeploymentEvent;
-import io.camunda.zeebe.client.api.response.PartitionInfo;
-import io.camunda.zeebe.client.api.response.Topology;
 import io.camunda.zeebe.it.clustering.ClusteringRule;
-import io.camunda.zeebe.model.bpmn.Bpmn;
 import io.camunda.zeebe.model.bpmn.BpmnModelInstance;
 import io.camunda.zeebe.model.bpmn.builder.ServiceTaskBuilder;
-import io.camunda.zeebe.protocol.record.Record;
-import io.camunda.zeebe.protocol.record.intent.CommandDistributionIntent;
-import io.camunda.zeebe.protocol.record.intent.DeploymentIntent;
-import io.camunda.zeebe.protocol.record.intent.JobIntent;
-import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.netty.util.NetUtil;
 import java.time.Duration;
 import java.util.List;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import org.junit.rules.ExternalResource;
 import org.slf4j.Logger;
 
@@ -41,6 +27,7 @@ public final class GrpcClientRule extends ExternalResource {
 
   protected ZeebeClient client;
   private final Consumer<ZeebeClientBuilder> configurator;
+  private ZeebeResourcesHelper resourcesHelper;
   private long startTime;
 
   public GrpcClientRule(final EmbeddedBrokerRule brokerRule) {
@@ -86,6 +73,7 @@ public final class GrpcClientRule extends ExternalResource {
         ZeebeClient.newClientBuilder().defaultRequestTimeout(Duration.ofSeconds(10));
     configurator.accept(builder);
     client = builder.build();
+    resourcesHelper = new ZeebeResourcesHelper(client);
     LOG.info("\n====\nClient startup time: {}\n====\n", (System.currentTimeMillis() - startTime));
     startTime = System.currentTimeMillis();
   }
@@ -98,6 +86,7 @@ public final class GrpcClientRule extends ExternalResource {
     client.close();
     LOG.info("Client closing time: " + (System.currentTimeMillis() - startTime));
     client = null;
+    resourcesHelper = null;
   }
 
   public ZeebeClient getClient() {
@@ -105,49 +94,28 @@ public final class GrpcClientRule extends ExternalResource {
   }
 
   public void waitUntilDeploymentIsDone(final long key) {
-    if (getPartitions().size() > 1) {
-      waitUntil(
-          () ->
-              RecordingExporter.commandDistributionRecords()
-                  .withDistributionIntent(DeploymentIntent.CREATE)
-                  .withRecordKey(key)
-                  .withIntent(CommandDistributionIntent.FINISHED)
-                  .exists());
-    } else {
-      waitUntil(
-          () ->
-              RecordingExporter.deploymentRecords()
-                  .withIntent(DeploymentIntent.CREATED)
-                  .withRecordKey(key)
-                  .exists());
-    }
+    resourcesHelper.waitUntilDeploymentIsDone(key);
   }
 
   public List<Integer> getPartitions() {
-    final Topology topology = client.newTopologyRequest().send().join();
-
-    return topology.getBrokers().stream()
-        .flatMap(i -> i.getPartitions().stream())
-        .filter(PartitionInfo::isLeader)
-        .map(PartitionInfo::getPartitionId)
-        .collect(Collectors.toList());
+    return resourcesHelper.getPartitions();
   }
 
   public long createSingleJob(final String type) {
-    return createSingleJob(type, b -> {}, "{}");
+    return resourcesHelper.createSingleJob(type);
   }
 
   public long createSingleJob(final String type, final Consumer<ServiceTaskBuilder> consumer) {
-    return createSingleJob(type, consumer, "{}");
+    return resourcesHelper.createSingleJob(type, consumer);
   }
 
   public long createSingleJob(
       final String type, final Consumer<ServiceTaskBuilder> consumer, final String variables) {
-    return createJobs(type, consumer, variables, 1).get(0);
+    return resourcesHelper.createSingleJob(type, consumer, variables);
   }
 
   public List<Long> createJobs(final String type, final int amount) {
-    return createJobs(type, b -> {}, "{}", amount);
+    return resourcesHelper.createJobs(type, amount);
   }
 
   public List<Long> createJobs(
@@ -156,69 +124,18 @@ public final class GrpcClientRule extends ExternalResource {
       final String variables,
       final int amount) {
 
-    final BpmnModelInstance modelInstance = createSingleJobModelInstance(type, consumer);
-    final long processDefinitionKey = deployProcess(modelInstance);
-
-    final var processInstanceKeys =
-        IntStream.range(0, amount)
-            .boxed()
-            .map(i -> createProcessInstance(processDefinitionKey, variables))
-            .collect(Collectors.toList());
-
-    final List<Long> jobKeys =
-        RecordingExporter.jobRecords(JobIntent.CREATED)
-            .withType(type)
-            .filter(r -> processInstanceKeys.contains(r.getValue().getProcessInstanceKey()))
-            .limit(amount)
-            .map(Record::getKey)
-            .collect(Collectors.toList());
-
-    assertThat(jobKeys).describedAs("Expected %d created jobs", amount).hasSize(amount);
-
-    return jobKeys;
-  }
-
-  public BpmnModelInstance createSingleJobModelInstance(
-      final String jobType, final Consumer<ServiceTaskBuilder> taskBuilderConsumer) {
-    return Bpmn.createExecutableProcess("process")
-        .startEvent("start")
-        .serviceTask(
-            "task",
-            t -> {
-              t.zeebeJobType(jobType);
-              taskBuilderConsumer.accept(t);
-            })
-        .endEvent("end")
-        .done();
+    return resourcesHelper.createJobs(type, consumer, variables, amount);
   }
 
   public long deployProcess(final BpmnModelInstance modelInstance) {
-    final DeploymentEvent deploymentEvent =
-        getClient()
-            .newDeployResourceCommand()
-            .addProcessModel(modelInstance, "process.bpmn")
-            .send()
-            .join();
-    waitUntilDeploymentIsDone(deploymentEvent.getKey());
-    return deploymentEvent.getProcesses().get(0).getProcessDefinitionKey();
+    return resourcesHelper.deployProcess(modelInstance);
   }
 
   public long createProcessInstance(final long processDefinitionKey, final String variables) {
-    return getClient()
-        .newCreateInstanceCommand()
-        .processDefinitionKey(processDefinitionKey)
-        .variables(variables)
-        .send()
-        .join()
-        .getProcessInstanceKey();
+    return resourcesHelper.createProcessInstance(processDefinitionKey, variables);
   }
 
   public long createProcessInstance(final long processDefinitionKey) {
-    return getClient()
-        .newCreateInstanceCommand()
-        .processDefinitionKey(processDefinitionKey)
-        .send()
-        .join()
-        .getProcessInstanceKey();
+    return resourcesHelper.createProcessInstance(processDefinitionKey);
   }
 }
